task: add resource filtering and new asset buttons
