+++ packages/client/hmi-client/src/page/project/tera-project.vue
@@ -15,6 +15,7 @@
 					@open-asset="openAssetFromSidebar"
 					@close-tab="removeClosedTab"
 					@remove-asset="removeAsset"
+					@open-new-asset="openNewAsset"
 				/>
 			</template>
 		</tera-slider-panel>
@@ -38,6 +39,7 @@
 					v-model:tabs="tabs"
 					@asset-loaded="setActiveTab"
 					@close-current-tab="removeClosedTab(activeTabIndex as number)"
+					@open-new-asset="openNewAsset"
 				/>
 			</SplitterPanel>
 			<SplitterPanel class="project-page top-z-index" v-if="workflowNode" :size="20">
@@ -121,6 +123,38 @@
 				<tera-notes-sidebar :asset-id="assetId" :page-type="pageType" />
 			</template>
 		</tera-slider-panel>
+		<!-- New model modal -->
+		<template>
+			<Teleport to="body">
+				<tera-modal
+					v-if="isNewModelModalVisible"
+					class="modal"
+					@modal-mask-clicked="isNewModelModalVisible = false"
+				>
+					<template #header>
+						<h4>New model</h4>
+					</template>
+					<template #default>
+						<form>
+							<label for="new-model">Enter a unique name for your model</label>
+							<InputText
+								v-bind:class="invalidInputStyle"
+								id="new-model"
+								type="text"
+								v-model="newModelName"
+								placeholder="new model"
+							/>
+						</form>
+					</template>
+					<template #footer>
+						<Button @click="createNewModel">Create model</Button>
+						<Button class="p-button-secondary" @click="isNewModelModalVisible = false">
+							Cancel
+						</Button>
+					</template>
+				</tera-modal>
+			</Teleport>
+		</template>
 	</main>
 </template>
 
@@ -150,7 +184,12 @@ import TeraSimulateCiemss from '@/components/workflow/tera-simulate-ciemss.vue';
 import TeraStratify from '@/components/workflow/tera-stratify.vue';
 import teraSimulateEnsembleCiemss from '@/components/workflow/tera-simulate-ensemble-ciemss.vue';
 import teraCalibrateEnsembleCiemss from '@/components/workflow/tera-calibrate-ensemble-ciemss.vue';
-import { workflowEventBus } from '@/services/workflow';
+import { createWorkflow, emptyWorkflow, workflowEventBus } from '@/services/workflow';
+import { newAMR } from '@/model-representation/petrinet/petrinet-service';
+import { createModel } from '@/services/model';
+import TeraModal from '@/components/widgets/tera-modal.vue';
+import Button from 'primevue/button';
+import InputText from 'primevue/inputtext';
 import TeraProjectPage from './components/tera-project-page.vue';
 
 // Asset props are extracted from route
@@ -175,6 +214,20 @@ workflowEventBus.on('drilldown', (payload: any) => {
 
 const isResourcesSliderOpen = ref(true);
 const isNotesSliderOpen = ref(false);
+
+// New Model Modal
+const isNewModelModalVisible = ref<boolean>(false);
+const newModelName = ref<string>('');
+const isValidName = ref<boolean>(true);
+const invalidInputStyle = computed(() => (!isValidName.value ? 'p-invalid' : ''));
+
+const existingModelNames = computed(() => {
+	const modelNames: string[] = [];
+	props.project.assets?.models.forEach((item) => {
+		modelNames.push(item.name);
+	});
+	return modelNames;
+});
 
 // Associated with tab storage
 const projectContext = computed(() => props.project?.id.toString());
@@ -239,6 +292,87 @@ async function removeAsset(asset: Tab) {
 	logger.error(`Failed to remove ${assetName}`, { showToast: true });
 }
 
+function createNewModel() {
+	if (newModelName.value.trim().length === 0) {
+		isValidName.value = false;
+		logger.info('Model name cannot be empty - please enter a different name');
+		return;
+	}
+	if (existingModelNames.value.includes(newModelName.value.trim())) {
+		isValidName.value = false;
+		logger.info('Duplicate model name - please enter a different name');
+		return;
+	}
+	isValidName.value = true;
+	newModel(newModelName.value.trim());
+	isNewModelModalVisible.value = false;
+}
+
+const newModel = async (modelName: string) => {
+	// 1. Load an empty AMR
+	const amr = newAMR(modelName);
+	(amr as any).id = undefined; // FIXME: id hack
+
+	const response = await createModel(amr);
+	const modelId = response?.id;
+
+	// 2. Add the model to the project
+	await ProjectService.addAsset(props.project.id, ProjectAssetTypes.MODELS, modelId);
+
+	// 3. Reroute
+	router.push({
+		name: RouteName.ProjectRoute,
+		params: {
+			assetName: 'Model',
+			pageType: ProjectAssetTypes.MODELS,
+			assetId: modelId
+		}
+	});
+};
+
+const openWorkflow = async () => {
+	// Create a new workflow
+	let wfName = 'workflow';
+	if (props.project && props.project.assets) {
+		wfName = `workflow ${props.project.assets[ProjectAssetTypes.SIMULATION_WORKFLOW].length + 1}`;
+	}
+	const wf = emptyWorkflow(wfName, '');
+
+	// FIXME: TDS bug thinks that k is z, June 2023
+	// @ts-ignore
+	wf.transform.z = 1;
+
+	// Add the workflow to the project
+	const response = await createWorkflow(wf);
+	const workflowId = response.id;
+	await ProjectService.addAsset(
+		props.project.id,
+		ProjectAssetTypes.SIMULATION_WORKFLOW,
+		workflowId
+	);
+
+	router.push({
+		name: RouteName.ProjectRoute,
+		params: {
+			assetName: 'Workflow',
+			pageType: ProjectAssetTypes.SIMULATION_WORKFLOW,
+			assetId: workflowId
+		}
+	});
+};
+
+const openNewAsset = (assetType: string) => {
+	switch (assetType) {
+		case ProjectAssetTypes.MODELS:
+			isNewModelModalVisible.value = true;
+			break;
+		case ProjectAssetTypes.SIMULATION_WORKFLOW:
+			openWorkflow();
+			break;
+		default:
+			break;
+	}
+};
 watch(
 	() => projectContext.value,
 	() => {
@@ -345,5 +479,9 @@ section,
 	display: inline-block;
 	overflow: hidden;
 	text-overflow: ellipsis;
+}
+
+.modal:deep(main) {
+	width: 50rem;
 }
 </style>
